Basic Encoding FaceRec
Based on A. Ng's implementation I created a similar implementation that can do simple face recognition. Most important next steps are to have different images for different people to allow for the software to create a virtual identity.

Another good idea is to also do face detection later.

This was good progress, scrum-style.